1) Added Size method: this implied that Fixed8 implements now the serializable interface. 2) Added few arithmetic operation (Add, Sub, div): this will be used to calculated networkfeeand feePerByte. Changed return value of the Value() method to int instead of int64. Modified fixed8_test accordingly.

+++ pkg/util/fixed8_test.go
@@ -13,7 +13,7 @@ func TestNewFixed8(t *testing.T) {
 
 	for _, val := range values {
 		assert.Equal(t, Fixed8(val*decimals), NewFixed8(val))
-		assert.Equal(t, int64(val), NewFixed8(val).Value())
+		assert.Equal(t, val, NewFixed8(val).Value())
 	}
 }
